+++ src/index.ts
@@ -1,3 +1,5 @@
 export * from './array/random';
 export * from './array/shuffle';
 export * from "./array/pickN";
+
+export * from './math/chance';